% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tika_jar.R
\name{tika_jar}
\alias{tika_jar}
\title{Path to Apache Tika}
\usage{
tika_jar()
}
\value{
A character vector of length one describing the path to the particular  \code{tika-app-X.XX.jar}.
}
\description{
A library for text and data extraction
}
\examples{
jar <- tika_jar()
# see help
sys::exec_wait('java',c('-jar',jar, '--help'))
# detect language of web page
sys::exec_wait('java',c('-jar',jar, '--language','https://tika.apache.org/'))
}
