% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tika_text.R
\name{tika_text}
\alias{tika_text}
\title{Extract plain text rendition}
\usage{
tika_text(input, ...)
}
\arguments{
\item{input}{Character vector describing the paths and/or urls to the input documents.}

\item{...}{Other parameters to be sent to `tika`.}
}
\value{
A character vector in the same order and with the same length as \code{input}. Unprocessed files are \code{as.character(NA)}.
}
\description{
Extract plain text rendition
}
\examples{
\donttest{
input= 'https://cran.r-project.org/doc/manuals/r-release/R-data.pdf'
output = tika_text(input)
cat(output)
}
}
