% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_tika.R
\name{install_tika}
\alias{install_tika}
\title{Install or Update the Apache Tika \code{jar}}
\usage{
install_tika(version = 1.18, md5_sum = "",
  mirrors = c("http://mirrors.ocf.berkeley.edu/apache/tika/",
  "http://apache.cs.utah.edu/tika/",
  "http://mirror.cc.columbia.edu/pub/software/apache/tika/"), retries = 2,
  url = "https://builds.apache.org/job/tika-branch-1x/10/org.apache.tika$tika-app/artifact/org.apache.tika/tika-app/1.18-20180309.175001-20/tika-app-1.18-20180309.175001-20.jar")
}
\arguments{
\item{version}{The declared Tika version}

\item{md5_sum}{The md5 checksum. Set to an empty string \code{""} to skip the check.}

\item{mirrors}{A vector Apache mirror sites. One is picked randomly.}

\item{retries}{The number of times to try the download.}

\item{url}{Optional url of a particular file to download. Will override downloading from random mirrors.}
}
\value{
Logical if the installation was successful.
}
\description{
This downloads and installs the Tika App \code{jar} into a user directory,
and verifies the integrity of the file using a md5 checksum.
The default settings should work fine.
}
\section{Details}{

The default settings of \code{install_tika()} should typically be left as they are.

This function will download the version of the Tika \code{jar} tested to work
with this package, and can verify file integrity using a checksum.

It will normally download from a random Apache mirror.
If the mirror fails,
it tries the archive at \code{http://archive.apache.org/dist/tika/}.
You can also enter a value for \code{url} directly to override this.

It will download into a directory determined
by the \code{rappdirs::user_data_dir()} function,
specific to the operating system.

If \code{tika()} is stopping with an error compalining about the \code{jar},
try running \code{install_tika()} again.
}

\section{Uninstalling}{

If you are uninstalling the entire \code{rtika} package
and want to remove the Tika App \code{jar} also,
before uninstalling run:

 \code{file.remove(tika_jar())}

Alternately, navigate to the install folder and delete it manually.
It is the file path returned by \code{rappdirs::user_data_dir('rtika')}.
The path is OS specific, and explained here:
https://github.com/r-lib/rappdirs .
}

\examples{
\dontrun{
install_tika()
}
}
