% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tika_check.R
\name{tika_check}
\alias{tika_check}
\title{Check tika_jar() against a md5 checksum}
\usage{
tika_check(md5_sum)
}
\arguments{
\item{md5_sum}{Character vector of length one with the target md5 checksum.}
}
\value{
logical if the .jar matches the md5
}
\description{
Checks the incuded Tika library, using the \code{tools::md5sum()} command.
The latest checksums are at https://tika.apache.org/download.html.
Note that these may ahead of the Tika version included in this package.
}
\examples{
tika_check('e2720c2392c1bd6634cc4a8801f7363a') #checksum of tika-app-1.17.jar
}
