% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tika_html.R
\name{tika_html}
\alias{tika_html}
\title{Extract html rendition}
\usage{
tika_html(input, ...)
}
\arguments{
\item{input}{Character vector describing the paths and/or urls to the input documents.}

\item{...}{Other parameters to be sent to `tika`.}
}
\value{
A character vector in the same order and with the same length as \code{input}. Unprocessed files are \code{as.character(NA)}.
}
\description{
Extract html rendition
}
\examples{
\donttest{
input= 'https://cran.r-project.org/doc/manuals/r-release/R-data.pdf'
output = tika_html(input)
}
}
