% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tika_xml.R
\name{tika_xml}
\alias{tika_xml}
\title{Extract xml rendition}
\usage{
tika_xml(input, ...)
}
\arguments{
\item{input}{Character vector describing the paths and/or urls to the input documents.}

\item{...}{Other parameters to be sent to \code{tika()}.}
}
\value{
A character vector in the same order and with the same length as \code{input}. Unprocessed files are \code{as.character(NA)}.
}
\description{
Extract xml rendition
}
\examples{
\donttest{
input= 'https://cran.r-project.org/doc/manuals/r-release/R-data.pdf'
output = tika_xml(input)
}
}
